+++ Sources/Client/Client/Client+WebSocket.swift
@@ -94,6 +94,19 @@ extension Client: WebSocketEventDelegate {
             userAtomic.set(user)
             return true
             
+        case .userUnbanned(let user, let cid, _, _):
+            if let cid = cid {
+                watchingChannelsAtomic.update { (channelsByCid) -> [ChannelId: [WeakRef<Channel>]] in
+                    var channelsByCid = channelsByCid
+                    if let weakChannels = channelsByCid[cid] {
+                        weakChannels.forEach({ $0.value?.bannedUsers.removeAll(where: { $0.id ==  user.id }) })
+                        channelsByCid[cid] = weakChannels
+                    }
+                    return channelsByCid
+                }
+            }
+            return true
+            
         case let .messageNew(message, _, _, _) where message.user != user && user.isMuted(user: message.user):
             // FIXIT: This shouldn't be by default.
             logger?.log("Skip a message (\(message.id)) from muted user (\(message.user.id)): \(message.textOrArgs)", level: .info)